+++ core/src/test/java/org/hisp/dhis/android/core/sms/TestRepositories.java
@@ -28,28 +28,34 @@
 
 package org.hisp.dhis.android.core.sms;
 
-import org.hisp.dhis.android.core.common.BaseDataModel;
 import org.hisp.dhis.android.core.common.State;
+import org.hisp.dhis.android.core.enrollment.Enrollment;
+import org.hisp.dhis.android.core.event.Event;
 import org.hisp.dhis.android.core.sms.domain.repository.LocalDbRepository;
 import org.hisp.dhis.android.core.sms.domain.repository.WebApiRepository;
+import org.hisp.dhis.android.core.trackedentity.TrackedEntityAttributeValue;
+import org.hisp.dhis.android.core.trackedentity.TrackedEntityInstance;
 import org.hisp.dhis.smscompression.models.Metadata;
 
+import java.util.ArrayList;
+import java.util.Collections;
 import java.util.Date;
+import java.util.List;
 
 import io.reactivex.Completable;
 import io.reactivex.Single;
 
 public class TestRepositories {
 
+    public static String enrollmentUid = "jQK0XnMVFIK";
+    public static String teiUid = "MmzaWDDruXW";
+    public static String trackedEntityType = "nEenWmSyUEp";
+
     public static class TestWebApiRepository implements WebApiRepository {
         public Metadata metadata;
 
         public TestWebApiRepository() {
-            this(new Metadata());
-        }
-
-        public TestWebApiRepository(Metadata metadata) {
-            this.metadata = metadata;
+            metadata = new TestMetadata();
         }
 
         @Override
@@ -63,14 +69,9 @@ public class TestRepositories {
         private String gatewayNumber = null;
         private String confirmationSenderNumber = null;
         private Integer resultWaitingTimeout = 120;
-        public Metadata metadata;
+        public Metadata metadata = new TestMetadata();
 
         public TestLocalDbRepository() {
-            this(new Metadata());
-        }
-
-        public TestLocalDbRepository(Metadata metadata) {
-            this.metadata = metadata;
             metadata.lastSyncDate = new Date();
         }
 
@@ -110,11 +111,6 @@ public class TestRepositories {
         }
 
         @Override
-        public Completable updateSubmissionState(BaseDataModel event, State sentViaSms) {
-            return Completable.complete();
-        }
-
-        @Override
         public Single<Metadata> getMetadataIds() {
             return Single.fromCallable(() -> metadata);
         }
@@ -122,6 +118,90 @@ public class TestRepositories {
         @Override
         public Completable setMetadataIds(Metadata metadata) {
             return Completable.complete();
+        }
+
+        @Override
+        public Single<Event> getEventToSubmit(String eventUid, String teiUid) {
+            return null;
+        }
+
+        @Override
+        public Single<TrackedEntityInstance> getTeiEnrollmentToSubmit(String enrollmentUid, String teiUid) {
+            return Single.just(TrackedEntityInstance.builder()
+                    .uid(teiUid)
+                    .trackedEntityType(trackedEntityType)
+                    .trackedEntityAttributeValues(getTestValues())
+                    .enrollments(Collections.singletonList(getTestEnrollment(enrollmentUid, teiUid)))
+                    .build());
+        }
+
+        @Override
+        public Completable updateEventSubmissionState(String eventUid, State state) {
+            return Completable.complete();
+        }
+
+        @Override
+        public Completable updateEnrollmentSubmissionState(String enrollmentUid, State state) {
+            return Completable.complete();
+        }
+    }
+
+    public static Enrollment getTestEnrollment(String enrollmentUid, String teiUid) {
+        return Enrollment.builder()
+                .uid(enrollmentUid)
+                .created(new Date())
+                .lastUpdated(new Date())
+                .organisationUnit("DiszpKrYNg8")
+                .program("IpHINAT79UW")
+                .enrollmentDate(new Date())
+                .trackedEntityInstance(teiUid)
+                .id(341L).build();
+    }
+
+    public static ArrayList<TrackedEntityAttributeValue> getTestValues() {
+        ArrayList<TrackedEntityAttributeValue> list = new ArrayList<>();
+        list.add(getTestValue("w75KJ2mc4zz", "Anne"));
+        list.add(getTestValue("zDhUuAYrxNC", "Anski"));
+        list.add(getTestValue("cejWyOfXge6", "Female"));
+        list.add(getTestValue("mLur0EGaw9A", "OU test"));
+        return list;
+    }
+
+    private static TrackedEntityAttributeValue getTestValue(String attr, String value) {
+        return TrackedEntityAttributeValue.builder()
+                .value(value)
+                .created(new Date())
+                .lastUpdated(new Date())
+                .trackedEntityAttribute(attr)
+                .trackedEntityInstance("MmzaWDDruXW")
+                .build();
+    }
+
+    public static class TestMetadata extends Metadata {
+        Enrollment enrollment = getTestEnrollment(enrollmentUid, teiUid);
+
+        public List<String> getUsers() {
+            return Collections.singletonList(TestRepositories.TestLocalDbRepository.userId);
+        }
+
+        public List<String> getTrackedEntityTypes() {
+            return Collections.singletonList(enrollment.trackedEntityInstance());
+        }
+
+        public List<String> getTrackedEntityAttributes() {
+            ArrayList<String> attrs = new ArrayList<>();
+            for (TrackedEntityAttributeValue item : getTestValues()) {
+                attrs.add(item.trackedEntityAttribute());
+            }
+            return attrs;
+        }
+
+        public List<String> getPrograms() {
+            return Collections.singletonList(enrollment.program());
+        }
+
+        public List<String> getOrganisationUnits() {
+            return Collections.singletonList(enrollment.organisationUnit());
         }
     }
 }
